requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
openai==1.3.0
flask==3.0.0
